Launch Heroku app
