+++ DS_Med_Cabinet/web_app/__init__.py
@@ -1,0 +1,19 @@
+#__init__.py
+
+
+
+from flask import Flask
+from Flask_API import Flask_API
+
+
+def create_app():
+    
+    app = Flask(__name__)
+    
+    app.register_blueprint(Flask_API)
+
+    return app
+
+if __name__ == "__main__":
+    my_app = create_app()
+    my_app.run(debug=True)
