+++ Pipfile
@@ -1,0 +1,17 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+pandas = "*"
+requests = "*"
+pickle-mixin = "*"
+Flask = "*"
+Flask-RESTful = "*"
+gunicorn = "*"
+
+[requires]
+python_version = "3.7"
