+++ Procfile
@@ -1,0 +1,1 @@
+web: gunicorn "web_app:create_app()"
